graph assigns nodes to computations, not hashes
Whereas the old graph had hash-keyed thunks, the new one has its thunks keyed
by opaque ids. This allows us to fully buy into the notion that thunks *stay in
place* once added to the graph.

This allows us to more easily track computation readiness for reduction,
because that computations is dependency-sensitive, and this means that our
dependencies are not changing all the time.